Resolve #14
The `XperienceLoggerProcessor` was removed because Xperience already has a `BufferedInMemoryEventLogWriter`, which works in conjunction with the default database writer. Internally it works with a `ConcurrentQueue` of its own. Thus, there is no need to double that in the processor. The buffered writer also gets explicitly called during the app initialization phase, when the database gets available. Thus, there is no need to check for `CMSApplication.ApplicationInitialized` in the logger.

OTOH, there is no point in trying to make the logger resilient (via Polly or others). It is a job of the product itself (which sadly does not have any resilience features as of yet).

The logger was made site-aware, as noted in https://github.com/Kentico/xperience-training-13/commit/0d3aebf87ce3dc9b41d08e175c944c9942555f15 .

TRAIN-1625, TRAIN-1626, TRAIN-1627